Move length_test.go into geom_test.go
